third_party/blink/web_tests/external/wpt/editing/run/fontname_2001-last-expected.txt

This is a testharness.js-based test.
Found 8 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "fo[o<listing>b]ar</listing>" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "fo[o<pre>b]ar</pre>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "fo[o<pre>b]ar</pre>" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "fo[o<pre>b]ar</pre>" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "fo[o<samp>b]ar</samp>" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "fo[o<samp>b]ar</samp>" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "fo[o<tt>b]ar</tt>" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "fo[o<tt>b]ar</tt>" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
Harness: the test ran to completion.
